fix ios target
